split out session into its own file
